fix: 상품 검색 결과 정합성 보완
