+++ docs/ia.md
@@ -24,7 +24,7 @@
 | Depth 1       | Depth 2        | URL                      | 설명                            | 권한   | 구현 상태       |
 | ------------- | -------------- | ------------------------ | ------------------------------- | ------ | --------------- |
 | 🏠 소비자 홈  | -              | `/`                      | 상품 목록, 필터, 검색           | 전체   | ✅ 구현됨       |
-| 🔍 검색       | -              | `/search`                | 상품/가게 검색 결과             | 전체   | 🚧 미구현 (T56) |
+| 🔍 검색       | -              | `/search`                | 상품 검색 결과                  | 전체   | ✅ 완료 (T56)   |
 | 📦 상품 상세  | -              | `/products/[productId]`  | 상품 정보, 가게 정보            | 전체   | ✅ 구현됨       |
 | 🛒 예약하기   | -              | `/order/[productId]`     | 수량 선택, 픽업 시간 확정       | 로그인 | ✅ 구현됨       |
 | 💳 결제       | -              | `/payment/toss-checkout` | Toss 결제 위젯                  | 로그인 | ✅ 구현됨       |
